fix selectPostListByUserId方法出参调整为Long

+++ ruoyi-system/src/main/java/com/ruoyi/system/mapper/SysPostMapper.java
@@ -18,7 +18,7 @@ public interface SysPostMapper extends BaseMapperPlus<SysPost> {
      * @param userId 用户ID
      * @return 选中岗位ID列表
      */
-    List<Integer> selectPostListByUserId(Long userId);
+    List<Long> selectPostListByUserId(Long userId);
 
     /**
      * 查询用户所属岗位组
